% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{boot.ci}
\alias{boot.ci}
\title{Bootstrap CI}
\usage{
boot.ci(x, B = 10000, conf = 0.95)
}
\arguments{
\item{x}{Vector of data}

\item{B}{Number of bootstrap replicates to run (defaults to 10000)}

\item{conf}{Confidence level (defaults to 0.95)}
}
\description{
Creates a bootstrap CI for the mean
}
\examples{
x <- rpois(10,10)
boot.ci(x)
}
\author{
Matt Tyers
}

