% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{caterprep.gls}
\alias{caterprep.gls}
\title{Only function is to make an input object for caterplot.gls()}
\usage{
caterprep.gls(mod)
}
\arguments{
\item{mod}{A gls model object from gls()}
}
\value{
An object to use in caterplot.gls()
}
\description{
Makes an input object for caterplot.gls().  This part of the algorithm takes some time, so better to do it all once, I say!
}
\author{
Matt Tyers
}

