% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{cdfplot}
\alias{cdfplot}
\title{CDF plot}
\usage{
cdfplot(a, b = NULL, c = NULL, d = NULL, e = NULL, lty = 0, lwd = 1,
  col = 1, xlab = "", ylab = "Cumulative Proportion", main = "",
  labels = NULL, yinfl = 1, showN = TRUE, ...)
}
\arguments{
\item{a}{Input vector}

\item{b}{Input vector}

\item{c}{Input vector}

\item{d}{Input vector}

\item{e}{Input vector}

\item{lty}{Line type.  Accepts one value or a vector of values for each line.  Specifying zero will automatically generate one line type for each line.}

\item{lwd}{Line width.  Accepts one value or a vector of values for each line.  Specifying zero will automatically generate one line width for each line.}

\item{col}{Line color.  Accepts one value or a vector of values for each line.  Specifying zero will automatically generate one line color for each line.}

\item{xlab}{Just like usual}

\item{ylab}{Just like usual}

\item{main}{Just like usual}

\item{labels}{A vector of labels for the legend.}

\item{yinfl}{Inflation factor for determining plotting window size.  Setting \code{yinfl=1} will use the maximum CDF value (1).  Setting it slightly higher allows more space for a legend.}

\item{showN}{Whether to include sample sizes in the legend.  Defaults to \code{TRUE}.}

\item{...}{Additional plotting arguments}
}
\description{
Produces an empirical CDF plot of up to five vectors (this can be biggerized)
}
\examples{
a <- rnorm(100)
b <- rnorm(100)
c <- rnorm(100,1,1)
d <- rnorm(100,0,2)
cdfplot(a,b,c,d,labels=c("apples","bananas","cucumbers","durians"),lty=1,col=1:4)
}
\author{
Matt Tyers
}

