% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{com.fac}
\alias{com.fac}
\title{Factor combiner thingy}
\usage{
com.fac(a, b, c = NULL, d = NULL, e = NULL, f = NULL)
}
\arguments{
\item{a}{Input vector}

\item{b}{Input vector}

\item{c}{Input vector}

\item{d}{Input vector}

\item{e}{Input vector}

\item{f}{Input vector}
}
\description{
Combines vectors of factors so you don't have to deal with specifying the levels
}
\examples{
a <- as.factor(c("Apples","Bananas","Apples","Bananas"))
b <- as.factor(c("Cucumbers","Bananas","Cucumbers","Durians"))
c <- com.fac(a,b)
summary(c)
}
\author{
Matt Tyers
}

