% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{dotsplot.rgb}
\alias{dotsplot.rgb}
\title{Color-ramp scatterplot}
\usage{
dotsplot.rgb(x, y, z, ramp = "gr", pch = 16, add = F, legend = T,
  cex = 1, neut = 0.5, legend.title = NULL, invert = F, xlab = "",
  ylab = "", main = "", ...)
}
\arguments{
\item{x}{X-coordinate vector}

\item{y}{Y-coordinate vector}

\item{z}{vector of values to use for the color ramp}

\item{ramp}{Type of color ramp to use, essentially possible ramps from red, green, and blue.  Possible values are "gr" (default), "rg", "gb", "bg", "br", "rb", "r", "g", "b", "rr", "gg", "bb"}

\item{pch}{Point character to use (defaults to 16, or solid circles)}

\item{add}{A value of TRUE adds the points to an active plot, rather than creating a new one}

\item{legend}{A value of FALSE suppresses the legend}

\item{cex}{Universal character expansion factor for plotting (between 0 and 1)}

\item{neut}{An RGB value (between 0 and 1) for the neutral color(s), those not used on the rgb ramp.  It might take a little twiddling.}

\item{legend.title}{Title for the legend}

\item{invert}{A value of TRUE inverts the current color ramp}

\item{...}{xlab, ylab, and main work like usual}
}
\description{
Creates a 2-d scatterplot with a third variable plotted on a color ramp
}
\examples{
x <- rnorm(1000)
y <- rnorm(1000)
z <- x+y
dotsplot.rgb(x,y,z,ramp="rg",legend.title="Legend Title")
dotsplot.rgb(x,y,z,ramp="rg",cex=.6,legend.title="Legend Title",invert=T)
}
\author{
Matt Tyers
}

