% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{getn}
\alias{getn}
\title{Get n}
\usage{
getn(x)
}
\arguments{
\item{x}{input vector}
}
\value{
Length (numeric)
}
\description{
Returns the length of a vector, excluding \code{NA} values
}
\examples{
x <- c(1,2,3,NA,5)
getn(x)
}
\author{
Matt Tyers
}

