% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{hhmm2dec}
\alias{hhmm2dec}
\title{Time format conversion}
\usage{
hhmm2dec(string, time = 24, out = "dd")
}
\arguments{
\item{string}{An input vector of times formatted as text (hh:mm)}

\item{time}{Value 24 or 12.  Specifies whether input is 24-hour time, or 12-hour + AM/PM}

\item{out}{Value "dd" or "dh".  Specifies whether output is in decimal days or decimal hours.}
}
\description{
pass in a character vector of times formatted hh:mm and it will output a vector of numeric times
}
\examples{
times <- c("11:00","12:15","22:30","0:15")
hhmm2dec(times,time=24,out="dd")
hhmm2dec(times,time=24,out="dh")

times12 <- c("11:00AM","12:15PM","2:30PM")
hhmm2dec(times12,time=12)
}
\author{
Matt Tyers
}

