% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{logit}
\alias{logit}
\title{Logit and Expit}
\usage{
logit(x)
}
\arguments{
\item{x}{Input}
}
\description{
Computes the logit log(x/(1/x)) or expit exp(x)/(1+exp(x))
}
\examples{
logit(0.42)
expit(0.42)

curve(expit(x),from=-5,to=5)
}
\author{
Matt Tyers
}

