% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{make.c}
\alias{make.c}
\title{Creates an assignment statement from a vector}
\usage{
make.c(x)
}
\arguments{
\item{x}{A vector of data}
}
\description{
Creates a c( , , ...) assignment statement from an existing vector.  Useful for copying data directly into BUGS, or hard-coding values into a routine.
}
\examples{
x <- sample(1:10,5)
make.c(x)
}
\author{
Matt Tyers
}

