% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{props}
\alias{props}
\title{Proportion plots and/or tables}
\usage{
props(x, y = rep(1, length(x)), numvar = 2,
  xname = deparse(substitute(x)), yname = deparse(substitute(y)),
  plot = T, kable = F, table = T)
}
\arguments{
\item{x}{First vector of factor data}

\item{y}{Second vector of factor data.  If a second vector is used, it makes plots/tables of the proportions of variable 1 for each unique value of variable 2}

\item{numvar}{Number of input variables (1 or 2)}

\item{xname}{Name of x to use for plots/tables}

\item{yname}{Name of y to use for plots/tables}

\item{plot}{Whether a plot is produced (TRUE or FALSE)}

\item{kable}{Whether to use kable format for tables, for use in R markdown (requires knitr package)
#'}

\item{table}{Whether a table is produced (TRUE or FALSE)}
}
\description{
Creates bar charts with 95\% Jeffries confidence intervals and/or output tables from input vectors of factor data
}
\examples{
lake <- sample(c("lake 1","lake 2"),20,replace=T)
spec <- sample(c("rainbow","cutthroat"),20,replace=T)
data.frame(lake,spec)

par(mfrow=c(2,1))
props(spec,lake,xname="Species",yname="Lake")
}
\author{
Matt Tyers
}

