% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{Simp}
\alias{Simp}
\title{Numeric integration using Simpson's Rule}
\usage{
Simp(f, a, b, m = 100)
}
\arguments{
\item{f}{An expression to be integrated as a function of t, e.g. "sin(2*pi*t)"}

\item{a}{Lower limit of integration}

\item{b}{Upper limit of integration}

\item{m}{Simpson's parameter (larger number is more accurate but slower)}
}
\description{
Numeric integration of a univariate function using Simpson's Rule.  NOTE: do NOT use variables with names f, a, b, m, n, h, sum, xx, or simp.
}
\examples{
Simp("t^2",0,1)
Simp("1.5+3*sin(2*pi*t)",0,0.0833333333,m=2000)
}
\author{
Matt Tyers
}

