% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MTfuncs_all.R
\name{tperm}
\alias{tperm}
\title{Permutation Alternative to 2-sample t-test}
\usage{
tperm(x1, x2, alternative = "2sided", reps = 10000)
}
\arguments{
\item{x1}{Data vector}

\item{x2}{Data vector}

\item{alternative}{The alternative hypothesis to test.  Options are \code{"2sided"} (the default),\code{"greater"}, or \code{"less"}.}

\item{reps}{The number of permutations to run.  Defaults to 10000.}
}
\value{
The permutation p-value (numeric)
}
\description{
Permutation Alternative to 2-sample 2-test, for inference to a difference in means.
}
\examples{
x1 <- rexp(100,3)
x2 <- rexp(100,2)
tperm(x1,x2)
}
\author{
Matt Tyers
}

